+++ libs/cta/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.0.6] - 2026-05-13
+
+### Changed
+
+- Update `@svta/cml-utils` to 1.5.0
+- Update `@svta/cml-structured-field-values` to 1.1.3
+
 ## [1.0.5] - 2026-02-11
 
 ### Changed
@@ -49,7 +56,8 @@ and this project adheres to
 - Convert to mono-repo ([#238](https://github.com/streaming-video-technology-alliance/common-media-library/issues/238))
 - Produce single bundled export for each package ([#260](https://github.com/streaming-video-technology-alliance/common-media-library/issues/260))
 
-[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.5...HEAD
+[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.6...HEAD
+[1.0.6]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.5...cta-v1.0.6
 [1.0.5]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.4...cta-v1.0.5
 [1.0.4]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.3...cta-v1.0.4
 [1.0.3]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.2...cta-v1.0.3